requirements.txt
gspread
oauth2client
flask
python-dotenv